Replace broken heuristics
Instead of blindly assuming inventory size based on game (which does not
work when a mod modifies inventory size), ask Minetest itself. When
inventory width is not specified by game (MTG), assume it to be 8.
Inventory height is reliably determined by looking at list length. If
the list length is not a multiple of inventory width, remaining items
are left untouched, avoiding any breakage.

Confirmed to work with MTG and MCL2.